Move conditional construction of AzureListener into a factory method, to consolidate Azure DevOps knowledge.
