+++ src/Fixie/Internal/Listeners/AzureListener.cs
@@ -10,6 +10,7 @@
     using System.Runtime.Versioning;
     using System.Text;
     using System.Threading;
+    using Cli;
     using Internal;
     using static System.Environment;
     using static Serialization;
@@ -37,6 +38,44 @@
         readonly int batchSize;
         readonly List<Result> batch;
         bool apiUnavailable;
+
+        internal static AzureListener Create()
+        {
+            if (ShouldUseAzureListener())
+                return new AzureListener();
+
+            return null;
+        }
+
+        static bool ShouldUseAzureListener()
+        {
+            var runningUnderAzure = GetEnvironmentVariable("TF_BUILD") == "True";
+
+            if (runningUnderAzure)
+            {
+                var accessTokenIsAvailable =
+                    !string.IsNullOrEmpty(GetEnvironmentVariable("SYSTEM_ACCESSTOKEN"));
+
+                if (accessTokenIsAvailable)
+                    return true;
+
+                using (Foreground.Yellow)
+                {
+                    WriteLine("The Azure DevOps access token has not been made available to this process, so");
+                    WriteLine("test results will not be collected. To resolve this issue, review your pipeline");
+                    WriteLine("definition to ensure that the access token is made available as the environment");
+                    WriteLine("variable SYSTEM_ACCESSTOKEN.");
+                    WriteLine();
+                    WriteLine("From https://docs.microsoft.com/en-us/azure/devops/pipelines/build/variables#systemaccesstoken");
+                    WriteLine();
+                    WriteLine("  env:");
+                    WriteLine("    SYSTEM_ACCESSTOKEN: $(System.AccessToken)");
+                    WriteLine();
+                }
+            }
+
+            return false;
+        }
 
         public AzureListener()
             : this(
